Added traversal of base types for function finding
Made the function finding traverse base types for components in addition to the component type. This gives access to stuff like LateUpdate on inherited classes where it previously may not have been shown. If you were missing some functions before when using EEE instead of the Unity event editor, this is probably why.